Add ical
ical is just a short loop for cal to make it nice for quick interactive
use. My use case is for quickly opening a terminal with ical with a
keybind.

The use of read -a and "${input[@]}" does *not* actually get proper
quoting semantics (try entering "foo bar" in ical), because read -a
breaks up the input using IFS without taking quoting into account.

I was choosing between either this or doing read -r input and expanding
an unquoted $input. Unquoted parameter expansion screams bad practice,
but this is actually a case where I *do* want word splitting, pathname
expansion, etc. to be applied. Can't do nested parameter expansion
though (type "$var" into ical and it will be passed literally to cal).

In the end I just went with one and rolled with it. Not like I expect to
pass more than just the usual date month year arguments with flags every
now and then. This is a good point for research though: what the best
practice for such a scenario is.