Add quicknote
I've found that sometimes I look for a quick throwaway editor window,
e.g. for scribbling down throwaway notes. This script is meant to
fulfill that purpose.

I have an i3 keybind that opens urxvt and execs this script, so as soon
as I quit the editor, the entire window quits.

Also noteworthy is that I change the CWD to ~/Documents/vim/quicknotes,
so that if I ever need to save any of my notes, it's simply a :wq
<filename> away.