Add mpd-notify
I use a small daemon called mpd-notification to automatically send
libnotify notifications whenever mpd plays a new song.

This is great, but with cmus I had a cmus-notify script that I could
trigger on-demand. Handy for e.g. quickly seeing the title of an
unfamiliar song while playing. As far as I can tell, mpd-notification
can't be triggered in this way.

So, I just quickly hacked out a script. The info is from mpc, formatted
the same way that mpd-notification does it.