Add rofi-mpc scripts
I grabbed a script from one of the issues for clerk (a rofi/fzf-based
frontend for mpd), which works quite simply by just listing songs and
playing the selected one.

rofi-mpc-playlist is restricted to the current playlist while
rofi-mpc-library grabs from the entire library and inserts the selection
into the playlist.