Add ez script
My ez shell function has been broken on zsh 5.1 for quite some time.
I've never really spent a lot of time tracking down what the issue is,
but it has to do with the interaction of fzf's tty-grabbing together
with zsh's <().

I've had a workaround on my Macbook for a while now: a binified bash
script, to work around the fact that this bug has to do with zsh. Might
as well check it in.