Add compositor-toggle
This is a toggling script in the same vein as locktoggle. This time, I
toggle a compositor, which I want for niceties when using Zoom, but
which I don't need any other time.

Because of this, I've elected to make it toggleable instead of starting
it unconditionally in my xinitrc, to save resources.