Improve notify scripts with dunstify
I now make use of dunstify in some notify scripts (in place of
notify-send), to allow notification IDs to be sent.

Notification IDs can be used to identify unique notifications, such that
e.g. the "Autolock on" and "Autolock off" notifications are treated as a
single notification being mutated, as opposed to numerous notifications
being displayed each time I trigger locktoggle.

I've also added hints (which display as a progress bar with dunst) and
font icons, to make the notifications look nicer -- particularly the
volume and brightness ones.