Add rnotify scripts
rnotify is a variant of notify-send, only this time it uses ssh to
trigger notify-send on a remote computer.

I've gotten a lot of mileage out of using it on a different machine
(e.g. my work machine, where I host and build all my work repos) in
order to notify me on my personal machine about finished scripts and
builds.