Use DPMS directly for auto-locking
Instead of relying on xautolock for locking the screen, I now directly
use DPMS, so both the screen dimming and the locking are governed by a
single mechanism.

In that regard, I've updated locktoggle to make use of a manual lockfile
for tracking toggle state, since now I can't use the existence of an
xautolock process for it.

This isn't perfect (I unconditionally set DPMS in my xinitrc on
boot-up), but it hasn't really been too much of a pain point. A simple
solution would be to track the lockfile in my xinitrc as well, but I'm
still mulling over whether I want such a dependency in my xinitrc.

This is in concert with a change across my dotfiles.